[CE-220] changes to enable debug log persistence
Utilize the 'debug' boolean attached to tasks
to trigger existing logic about creation, uploading,
and deletion of log files (that host log messages from
the debug level upward).

This allows service mode connectors built with this change
(and C1 built with [CE-221]) that are in 'debug' mode
(as currently set in the support dashboard for 24 hours
at a time) to keep log files for full sync tasks.
Log files are rotated after each full sync task,
and they are only uploaded for 'debug' full sync tasks.
It is not necessary to set `log-level` or any other
configuration at the connector / baton-runner level
to enable this behavior; in fact that setting has no
bearing on our logging here.

There is an existing skeleton in baton-sdk about enabling
debug logs at the Manager level - while this is feasible,
it likely requires a new Task (or different communication
structure) and we do not attempt to utilize this.
We do take care to leave it in roughly the same state it was.

+++ pkg/tasks/c1api/full_sync.go
@@ -145,6 +145,7 @@ func (c *fullSyncTaskHandler) HandleTask(ctx context.Context) error {
 	c1zPath := assetFile.Name()
 	err = assetFile.Close()
 	if err != nil {
+		l.Error("failed to close asset file", zap.Error(err))
 		return c.helpers.FinishTask(ctx, nil, nil, err)
 	}
 
@@ -183,8 +184,9 @@ func (c *fullSyncTaskHandler) HandleTask(ctx context.Context) error {
 		return c.helpers.FinishTask(ctx, nil, nil, err)
 	}
 
-	err = uploadDebugLogs(ctx, c.helpers)
+	err = uploadDebugLogs(ctx, c.helpers, c.task.GetDebug())
 	if err != nil {
+		l.Error("failed to upload debug Task logs", zap.Error(err))
 		return c.helpers.FinishTask(ctx, nil, nil, err)
 	}
 
@@ -211,7 +213,9 @@ func newFullSyncTaskHandler(
 	}
 }
 
-func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
+// Check if Debug logs should be uploaded to C1 and if so do so,
+// otherwise silently return success.
+func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers, deleteDebugLogs bool) error {
 	ctx, span := tracer.Start(ctx, "uploadDebugLogs")
 	defer span.End()
 
@@ -248,19 +252,27 @@ func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
 
 	debugfile, err := os.Open(debugPath)
 	if err != nil {
+		l.Error("failed to open debug log file path", zap.Error(err))
 		return err
 	}
-	defer func() {
-		err := os.Remove(debugPath)
-		if err != nil {
-			l.Error("failed to delete file with debug logs", zap.Error(err), zap.String("file", debugPath))
-		}
-	}()
+	if deleteDebugLogs {
+		// We only delete the debug log when asked to,
+		// as Manager-required log files are not automatically rotated.
+		defer func() {
+			err := os.Remove(debugPath)
+			if err != nil {
+				l.Error("failed to delete file with debug logs", zap.Error(err), zap.String("file", debugPath))
+			} else {
+				l.Info("deleted debug path")
+			}
+		}()
+	}
 	defer debugfile.Close()
 
 	l.Info("uploading debug logs", zap.String("file", debugPath))
 	err = helper.Upload(ctx, debugfile)
 	if err != nil {
+		l.Error("failed to upload debug logs", zap.Error(err))
 		return err
 	}
